+++ src/components/ui/dialog.tsx
@@ -126,7 +126,10 @@ function DialogTitle({
   return (
     <DialogPrimitive.Title
       data-slot="dialog-title"
-      className={cn("font-heading text-lg leading-none font-semibold", className)}
+      className={cn(
+        "font-heading text-lg leading-none font-semibold",
+        className
+      )}
       {...props}
     />
   );